refactor utils.js: split into render.js and common.js
