+++ src/components/film-card.js
@@ -1,4 +1,4 @@
-import {formatDuration} from "../utils.js";
+import {formatDuration} from "../utils/common.js";
 import AbstractComponent from "./abstract-component.js";
 
 const createFilmCardTemplate = (film) => {
