+++ src/main.js
@@ -7,7 +7,8 @@ import ShowMoreButton from "./components/show-more-button.js";
 import {generateFilms} from "./mock/film.js";
 import {generateUser} from "./mock/user.js";
 import FilmDetails from "./components/film-details.js";
-import {getRandomIntegerNumber, sortFilms, render, RenderPosition} from "./utils.js";
+import {getRandomIntegerNumber, sortFilms} from "./utils/common.js";
+import {render, RenderPosition} from "./utils/render.js";
 import Statistics from "./components/statistics.js";
 import SortMenu from "./components/sort-menu.js";
 import MainContentNoData from "./components/main-content-no-data.js";
