+++ src/utils/common.js
@@ -1,9 +1,4 @@
-import {months} from "./const";
-
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
+import {months} from "../const.js";
 
 const shuffle = (array) => {
   let j = 0;
@@ -68,22 +63,4 @@ export const getRandomArrayItem = (array) => {
 
 export const sortFilms = (field) => {
   return (a, b) => a[field] < b[field] ? 1 : -1;
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
 };
